Add being able to show statistics block in the flexible landing page
