+++ rca/project_styleguide/templates/patterns/pages/standardpages/flexible_landing_page.html
@@ -57,7 +57,9 @@
         <div class="page__content">
             {% for item in processed_body %}
                 {% with section=item.block %}
-                    {% if section.block_type == "promo_banner" %}
+                    {% if section.block_type == "statistics" %}
+                        {% include "patterns/organisms/stat-block/stat-block.html" with modifier="transparent" stats_block=section.value top_notch_bg=item.previous_block_bg bottom_notch_bg=item.next_block_bg %} 
+                    {% elif section.block_type == "promo_banner" %}
                         {% include_block section with previous_section_bg=item.previous_block_bg next_section_bg=item.next_block_bg %}
                     {% else %}
                         {# This is a body section block. #}
