+++ rca/standardpages/models.py
@@ -9,7 +9,7 @@ from wagtail.search import index
 from rca.standardpages.blocks import LandingPageBodyBlock
 from rca.utils.blocks import StoryBlock
 from rca.utils.fields import StreamField as CustomStreamField
-from rca.utils.models import BasePage, RelatedPage
+from rca.utils.models import BasePage, ProcessedBodyMixin, RelatedPage
 
 
 class InformationPageRelatedPage(RelatedPage):
@@ -64,7 +64,7 @@ class IndexPage(BasePage):
         return context
 
 
-class FlexibleLandingPage(BasePage):
+class FlexibleLandingPage(ProcessedBodyMixin, BasePage):
     template = "patterns/pages/standardpages/flexible_landing_page.html"
 
     hero_image = models.ForeignKey(
@@ -85,52 +85,6 @@ class FlexibleLandingPage(BasePage):
         FieldPanel("introduction"),
         FieldPanel("body"),
     ]
-
-    def get_processed_body(self):
-        # Processes the body streamfield to determine when and what notches are displayed.
-        processed_body = []
-        num_blocks = len(self.body)
-
-        for i, block in enumerate(self.body):
-            processed_section = {
-                "block": block,
-            }
-
-            previous_block = self.body[i - 1] if i > 0 else None
-            next_block = self.body[i + 1] if (i + 1) < num_blocks else None
-
-            is_last_block = next_block is None
-            next_is_promo_banner = next_block and next_block.block_type in [
-                "promo_banner"
-            ]
-            backgrounds_match = next_block and next_block.value.get(
-                "background_color"
-            ) == block.value.get("background_color")
-
-            # Don't display a notch in this section if:
-            # - This is the last block in the body.
-            # - The next block has the same background color as the current block.
-            processed_section["should_display_notch"] = not (
-                is_last_block or next_is_promo_banner or backgrounds_match
-            )
-
-            # If the block is a promo banner block, we need to check the
-            # previous and next block's background color to determine the background colors
-            # for the notch.
-            if block.block_type in ["promo_banner"]:
-                if previous_block and previous_block.block_type == "body_section":
-                    processed_section["previous_block_bg"] = previous_block.value.get(
-                        "background_color"
-                    )
-
-                if next_block and next_block.block_type == "body_section":
-                    processed_section["next_block_bg"] = next_block.value.get(
-                        "background_color"
-                    )
-
-            processed_body.append(processed_section)
-
-        return processed_body
 
     def get_context(self, request, *args, **kwargs):
         context = super().get_context(request, *args, **kwargs)
