+++ rca/utils/models.py
@@ -1099,6 +1099,61 @@ def get_listing_image(page):
     return image
 
 
+class ProcessedBodyMixin:
+    """Mixin for pages with a body StreamField containing sections separated by notches.
+
+    Currently used in HomePage and FlexibleLandingPage.
+    """
+
+    special_block_types = ["statistics", "promo_banner"]
+
+    def get_processed_body(self):
+        # Processes the body streamfield to determine when and what notches are displayed.
+        processed_body = []
+        num_blocks = len(self.body)
+
+        for i, block in enumerate(self.body):
+            processed_section = {
+                "block": block,
+            }
+
+            previous_block = self.body[i - 1] if i > 0 else None
+            next_block = self.body[i + 1] if (i + 1) < num_blocks else None
+
+            is_last_block = next_block is None
+            next_is_special = (
+                next_block and next_block.block_type in self.special_block_types
+            )
+            backgrounds_match = next_block and next_block.value.get(
+                "background_color"
+            ) == block.value.get("background_color")
+
+            # Don't display a notch in this section if:
+            # - This is the last block in the body.
+            # - The next block is a special block (e.g. statistics, promo_banner).
+            # - The next block has the same background color as the current block.
+            processed_section["should_display_notch"] = not (
+                is_last_block or next_is_special or backgrounds_match
+            )
+
+            # If the block is a special block, we need to check the previous and next
+            # block's background color to determine the background colors for the notch.
+            if block.block_type in self.special_block_types:
+                if previous_block and previous_block.block_type == "body_section":
+                    processed_section["previous_block_bg"] = previous_block.value.get(
+                        "background_color"
+                    )
+
+                if next_block and next_block.block_type == "body_section":
+                    processed_section["next_block_bg"] = next_block.value.get(
+                        "background_color"
+                    )
+
+            processed_body.append(processed_section)
+
+        return processed_body
+
+
 class TapMixin(models.Model):
     tap_widget = models.ForeignKey(
         "utils.TapWidgetSnippet",
